requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
# Note: openai SDK (already listed) doubles as Groq client via base_url override
httpx>=0.27.2
openai>=1.51.0
anthropic>=0.39.0
google-genai>=1.0.0
python-dotenv>=1.0.1
loguru>=0.7.2
redis>=5.1.1
